projects/01-ai-chatbot-cli/pyproject.toml
[project]
name = "01-ai-chatbot-cli"
version = "0.1.0"
description = "Learn about using openai api, managing chat history"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.82.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tiktoken>=0.5.0",
]
